# api/swagger.yaml
basePath: /api/v1
definitions:
...
  common.ErrorResponse:
    description: Standard error response
    properties:
      message:
        description: Human-readable error message
        example: 'invalid input body: field email is required'
        type: string
    type: object
...
  user.AddPassengerResponse:
    properties:
      passenger_id:
        description: Unique identifier of the created passenger
        example: 10
        type: integer
    type: object
  user.LoginRequest:
    properties:
      email:
        description: Email address of the user
        example: user@example.com
        type: string
      password:
        description: Account password
        example: securepass123
        type: string
    required:
    - email
    - password
    type: object
  user.LoginResponse:
    properties:
      token:
        description: JWT access token
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      user_id:
        description: Unique identifier of the authenticated user
        example: 1
        type: integer
    type: object
  user.PassengerInput:
    properties:
      birth_date:
        description: Date of birth in YYYY-MM-DD format
        example: "1990-05-20"
        type: string
      citizenship:
        description: Country of citizenship (ISO 3166-1 alpha-3)
        example: RUS
        type: string
      document_number:
        description: Travel document number
        example: "7712345678"
        type: string
      document_type:
        description: Travel document type
        example: passport
        type: string
      first_name:
        description: Legal first name as in travel document
        example: Ivan
        type: string
      gender:
        description: 'Gender: male or female'
        enum:
        - male
        - female
        example: male
        type: string
      last_name:
        description: Legal last name as in travel document
        example: Ivanov
        type: string
      middle_name:
        description: Middle name (patronymic), optional
        example: Petrovich
        type: string
    required:
    - birth_date
    - citizenship
    - document_number
    - document_type
    - first_name
    - gender
    - last_name
    type: object
  user.PassengerResponse:
    properties:
      birth_date:
        description: Date of birth in YYYY-MM-DD format
        example: "1990-05-20"
        type: string
      citizenship:
        description: Country of citizenship (ISO 3166-1 alpha-3)
        example: RUS
        type: string
      document_number:
        description: Travel document number
        example: "7712345678"
        type: string
      document_type:
        description: Travel document type
        example: passport
        type: string
      first_name:
        description: Legal first name
        example: Ivan
        type: string
      gender:
        description: 'Gender: male, female'
        enum:
        - male
        - female
        example: male
        type: string
      id:
        description: Unique passenger identifier
        example: 10
        type: integer
      last_name:
        description: Legal last name
        example: Ivanov
        type: string
      middle_name:
        description: Middle name (patronymic)
        example: Petrovich
        type: string
      user_id:
        description: Owner user identifier
        example: 1
        type: integer
    type: object
  user.RegisterRequest:
    properties:
      email:
        description: Email address of the new user
        example: user@example.com
        type: string
      password:
        description: Password must be at least 8 characters
        example: securepass123
        minLength: 8
        type: string
      phone_number:
        description: Phone number in international format
        example: "+79991234567"
        type: string
    required:
    - email
    - password
    - phone_number
    type: object
  user.RegisterResponse:
    properties:
      user_id:
        description: Unique identifier of the created user
        example: 1
        type: integer
    type: object
  user.UserResponse:
    properties:
      created_at:
        description: Account creation timestamp in RFC 3339 format
        example: "2025-01-15T10:30:00Z"
        type: string
      email:
        description: Email address
        example: user@example.com
        type: string
      id:
        description: Unique user identifier
        example: 1
        type: integer
      phone_number:
        description: Phone number
        example: "+79991234567"
        type: string
      role:
        description: 'User role: user or admin'
        enum:
        - user
        - ' admin'
        example: user
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Gateway API for flight booking service
  title: Flyte API
  version: "1.0"
paths:
...
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Authenticates a user by email and password, returns a JWT token.
      parameters:
      - description: Login credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authenticated successfully
          schema:
            $ref: '#/definitions/user.LoginResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Authenticate user
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Creates a new user account with the provided email, password, and
        phone number.
      parameters:
      - description: Registration data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/user.RegisterResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "409":
          description: Email already taken
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
  /users/{id}:
    get:
      description: Returns profile information for the specified user. Only the account
        owner can access this.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/user.UserResponse'
        "403":
          description: Forbidden — not the account owner
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - users
  /users/{id}/passengers:
    get:
      description: Returns all passengers belonging to the specified user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of passengers
          schema:
            items:
              $ref: '#/definitions/user.PassengerResponse'
            type: array
        "403":
          description: Forbidden — not the account owner
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List passengers
      tags:
      - passengers
    post:
      consumes:
      - application/json
      description: Creates a new passenger record linked to the specified user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Passenger data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.PassengerInput'
      produces:
      - application/json
      responses:
        "201":
          description: Passenger created
          schema:
            $ref: '#/definitions/user.AddPassengerResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden — not the account owner
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "409":
          description: Duplicate passenger document
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add a passenger
      tags:
      - passengers
  /users/{id}/passengers/{passengerId}:
    delete:
      description: Removes a passenger record belonging to the specified user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Passenger ID
        in: path
        name: passengerId
        required: true
        type: integer
      responses:
        "204":
          description: Passenger deleted
        "403":
          description: Forbidden — not the account owner
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Passenger not found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a passenger
      tags:
      - passengers
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"